refactor(my-diary): decompose app pieces
affects: @my-diary/app, @components/form-add-post, @components/post-controls, @components/post-list,
@practice/my-diary, @store/filter, @store/search-value, @ui/list

ISSUES CLOSED: #82
