+++ practice/my-diary/app/package.json
@@ -12,6 +12,11 @@
     "react-dom": "^17.0.1",
     "react-intl": "^5.15.7",
     "@emotion/styled": "^11.1.5",
+    "@components/form-add-post": "0.0.0",
+    "@components/post-controls": "0.0.0",
+    "@components/post-list": "0.0.0",
+    "@store/filter": "0.0.0",
+    "@store/search-value": "0.0.0",
     "@ui/button": "0.0.0",
     "@ui/icons": "0.0.0",
     "@ui/text": "0.0.0",
