+++ practice/my-diary/components/form-add-post/src/actions/index.ts
@@ -1,0 +1,1 @@
+export * from './addItem'
