+++ practice/my-diary/components/post-controls/src/index.ts
@@ -1,0 +1,1 @@
+export * from './PostControls'
