+++ practice/my-diary/components/post-list/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@components/post-list",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.ts",
+  "dependencies": {
+    "@emotion/react": "^11.1.5",
+    "@emotion/styled": "^11.1.5",
+    "react-intl": "^5.15.7",
+    "@store/filter": "0.0.0",
+    "@store/search-value": "0.0.0",
+    "@ui/layout": "^0.0.0",
+    "@ui/list": "^0.0.0",
+    "@ui/text": "^0.0.0",
+    "react": "^17.0.1",
+    "@store/notes": "0.0.0"
+  }
+}
