+++ practice/my-diary/components/post-list/src/actions/deleteItem.ts
