+++ practice/my-diary/components/post-list/src/index.ts
@@ -1,0 +1,1 @@
+export * from './PostList'
