+++ practice/my-diary/store/filter/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@store/filter",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.ts",
+  "devDependencies": {
+    "react": "^17.0.1"
+  }
+}
