+++ practice/my-diary/store/filter/src/useFilter.ts
@@ -1,0 +1,5 @@
+import { useContext } from 'react'
+
+import { Context }    from './Context'
+
+export const useFilter = () => useContext(Context)
