+++ practice/my-diary/store/search-value/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@store/search-value",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.ts",
+  "devDependencies": {
+    "react": "^17.0.1"
+  }
+}
