+++ practice/my-diary/store/search-value/src/Context.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+export const Context = createContext(null)
+export const { Provider } = Context
+export const { Consumer } = Context
