+++ practice/my-diary/store/search-value/src/useSearchValue.ts
@@ -1,0 +1,5 @@
+import { useContext } from 'react'
+
+import { Context }    from './Context'
+
+export const useSearchValue = () => useContext(Context)
