+++ practice/my-diary/ui/list/src/Item.tsx
@@ -6,7 +6,6 @@ import { Button }                         from '@ui/button'
 import { HeartIcon, StarIcon, TrashIcon } from '@ui/icons'
 import { Box, Layout, Row }               from '@ui/layout'
 import { Text }                           from '@ui/text'
-import { useNotes }                       from '@store/notes'
 
 const StyledItem = styled.li(
   () => ({
@@ -20,42 +19,39 @@ const StyledItem = styled.li(
   border,
 )
 
-export const Item = ({ note, deleteItem, toggleStatus }) => {
-  const [notes, setNotes] = useNotes()
-  return (
-    <>
-      <Row>
-        <StyledItem onDoubleClick={() => setNotes(toggleStatus(notes, note.id, 'liked'))}>
-          <Text lineHeight='40px' padding='0 6px'>
-            {note.note}
-          </Text>
-        </StyledItem>
+export const Item = ({ notes, setNotes, note, deleteItem, toggleStatus }) => (
+  <>
+    <Row>
+      <StyledItem onDoubleClick={() => setNotes(toggleStatus(notes, note.id, 'liked'))}>
+        <Text lineHeight='40px' padding='0 6px'>
+          {note.note}
+        </Text>
+      </StyledItem>
+      <Layout flexBasis={3} />
+      <Box>
+        <Button
+          backgroundColor='#fff'
+          border='none'
+          width='40px'
+          onClick={() => setNotes(toggleStatus(notes, note.id, 'important'))}
+        >
+          <StarIcon color={note.important ? '#FFD700' : '#6c757d'} />
+        </Button>
         <Layout flexBasis={3} />
-        <Box>
-          <Button
-            backgroundColor='#fff'
-            border='none'
-            width='40px'
-            onClick={() => setNotes(toggleStatus(notes, note.id, 'important'))}
-          >
-            <StarIcon color={note.important ? '#FFD700' : '#6c757d'} />
-          </Button>
-          <Layout flexBasis={3} />
-          <Button
-            backgroundColor='#fff'
-            border='none'
-            width='40px'
-            onClick={() => setNotes(deleteItem(notes, note.id))}
-          >
-            <TrashIcon />
-          </Button>
-          <Layout flexBasis={3} />
-          <Button backgroundColor='#fff' border='none' width='40px'>
-            <HeartIcon opacity={note.liked ? '1' : '0'} />
-          </Button>
-        </Box>
-      </Row>
-      <Layout flexBasis={6} />
-    </>
-  )
-}
+        <Button
+          backgroundColor='#fff'
+          border='none'
+          width='40px'
+          onClick={() => setNotes(deleteItem(notes, note.id))}
+        >
+          <TrashIcon />
+        </Button>
+        <Layout flexBasis={3} />
+        <Button backgroundColor='#fff' border='none' width='40px'>
+          <HeartIcon opacity={note.liked ? '1' : '0'} />
+        </Button>
+      </Box>
+    </Row>
+    <Layout flexBasis={6} />
+  </>
+)
